Separate caching to its own service
https://github.com/tipsy/profile-summary-for-github/issues/110
