+++ src/main/kotlin/app/UserService.kt
@@ -1,24 +1,9 @@
 package app
 
 import app.util.CommitCountUtil
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.JsonNode
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer
-import com.fasterxml.jackson.databind.module.SimpleModule
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import org.eclipse.egit.github.core.Repository
 import org.eclipse.egit.github.core.RepositoryCommit
-import org.eclipse.egit.github.core.User
-import org.eclipse.egit.github.core.UserPlan
 import org.slf4j.LoggerFactory
-import java.io.Serializable
-import java.sql.DriverManager
-import java.time.Instant
-import java.time.LocalDateTime
-import java.time.temporal.ChronoUnit
-import java.util.Date
 import java.util.concurrent.ConcurrentHashMap
 import java.util.stream.IntStream
 import kotlin.streams.toList
@@ -26,7 +11,6 @@ import kotlin.streams.toList
 object UserService {
 
     private const val pageSize = 100
-    private const val urlToDb = "jdbc:h2:mem:userinfo"
     private val log = LoggerFactory.getLogger("app.UserCtrlKt")
     private val repo = GhService.repos.getRepository("tipsy", "profile-summary-for-github")
     private val watchers = ConcurrentHashMap.newKeySet<String>()
@@ -42,13 +26,13 @@ object UserService {
         val remainingRequests by lazy { GhService.remainingRequests }
         val hasFreeRemainingRequests by lazy { remainingRequests > (freeRequestCutoff ?: remainingRequests) }
         return Config.unrestricted()
-                || (lookUpInCache(user) != null)
+                || (CacheService.lookUpInCache(user) != null)
                 || hasFreeRemainingRequests
                 || (remainingRequests > 0 && hasStarredRepo(user))
     }
 
     fun getUserProfile(username: String): UserProfile {
-        return (lookUpInCache(username) ?: generateUserProfile(username))
+        return (CacheService.lookUpInCache(username) ?: generateUserProfile(username))
     }
 
     private fun hasStarredRepo(username: String): Boolean {
@@ -82,78 +66,6 @@ object UserService {
         listOf()
     }
 
-    private fun createTableIfAbsent() {
-        val connection = DriverManager.getConnection(urlToDb)
-        val statement = connection.createStatement()
-
-        statement.execute(
-        "CREATE TABLE IF NOT EXISTS userinfo (" +
-            "id VARCHAR2 PRIMARY KEY," +
-            "timestamp TIMESTAMP WITH TIME ZONE, " +
-            "data JSON" +
-            ")"
-        )
-    }
-
-    private fun lookUpInCache(username: String): UserProfile? {
-        val connection = DriverManager.getConnection(urlToDb)
-
-        createTableIfAbsent()
-
-        val preparedStatement = connection.prepareStatement(
-        "SELECT " +
-            "timestamp, " +
-            "data " +
-            "FROM userinfo " +
-            "WHERE id = ?"
-        )
-        preparedStatement.setString(1, username)
-
-        val result = preparedStatement.executeQuery()
-        result.use {
-            // guaranteed to be at most one.
-            if (it.next()) {
-                val timestamp = it.getTimestamp(1).toLocalDateTime()
-                val diffInHours = ChronoUnit.HOURS.between(timestamp, LocalDateTime.now())
-                if (diffInHours <= 6) {
-                    val json: String = it.getString(2)
-
-                    log.debug("cache hit: {}", json)
-
-                    val simpleModule = SimpleModule()
-                    simpleModule.addDeserializer(UserProfile::class.java, UserProfile.Deserializer())
-
-                    val objectMapper = jacksonObjectMapper()
-                    objectMapper.registerModule(simpleModule)
-
-                    return objectMapper.readValue<UserProfile>(json)
-                }
-            }
-        }
-
-        log.debug("cache miss for username: {}", username)
-
-        return null
-    }
-
-    private fun saveInCache(userProfile: UserProfile) {
-        val connection = DriverManager.getConnection(urlToDb)
-
-        createTableIfAbsent()
-
-        val json = jacksonObjectMapper().writeValueAsString(userProfile)
-
-        val preparedStatement = connection.prepareStatement(
-        "MERGE INTO userinfo (id, timestamp, data) KEY (id) " +
-            "VALUES (?, CURRENT_TIMESTAMP(), ? FORMAT JSON)"
-        )
-
-        preparedStatement.setString(1, userProfile.user.login)
-        preparedStatement.setString(2, json)
-
-        preparedStatement.execute()
-    }
-
     private fun generateUserProfile(username: String): UserProfile {
         val user = GhService.users.getUser(username)
         val repos = GhService.repos.getRepositories(username).filter { !it.isFork && it.size != 0 }
@@ -182,112 +94,8 @@ object UserService {
             repoStarCountDescriptions
         )
 
-        saveInCache(userProfile)
+        CacheService.saveInCache(userProfile)
 
         return userProfile;
-    }
-}
-
-data class UserProfile(
-        val user: User,
-        val quarterCommitCount: Map<String, Int>,
-        val langRepoCount: Map<String, Int>,
-        val langStarCount: Map<String, Int>,
-        val langCommitCount: Map<String, Int>,
-        val repoCommitCount: Map<String, Int>,
-        val repoStarCount: Map<String, Int>,
-        val repoCommitCountDescriptions: Map<String, String?>,
-        val repoStarCountDescriptions: Map<String, String?>
-) : Serializable {
-    val timeStamp = Instant.now().toEpochMilli()
-
-    class Deserializer(valueClass: Class<*>?): StdDeserializer<UserProfile>(valueClass) {
-        constructor(): this(null)
-
-        override fun deserialize(jsonParser: JsonParser?, context: DeserializationContext?): UserProfile {
-            if (jsonParser == null || context == null) {
-                throw RuntimeException()
-            }
-
-            val userProfileNode: JsonNode = jsonParser.codec?.readTree(jsonParser) ?: throw RuntimeException()
-
-            val userNode = userProfileNode.get("user")
-            val user = User()
-            user.isHireable = userNode.booleanValue()
-            user.createdAt = Date(userNode.longValue())
-            user.collaborators = userNode.intValue()
-            user.diskUsage = userNode.intValue()
-            user.followers = userNode.intValue()
-            user.following = userNode.intValue()
-            user.id = userNode.intValue()
-            user.ownedPrivateRepos = userNode.intValue()
-            user.privateGists = userNode.intValue()
-            user.publicGists = userNode.intValue()
-            user.publicRepos = userNode.intValue()
-            user.totalPrivateRepos = userNode.intValue()
-            user.avatarUrl = userNode.textValue()
-            user.blog = userNode.textValue()
-            user.company = userNode.textValue()
-            user.email = userNode.textValue()
-            user.gravatarId = userNode.textValue()
-            user.htmlUrl = userNode.textValue()
-            user.location = userNode.textValue()
-            user.login = userNode.textValue()
-            user.name = userNode.textValue()
-            user.type = userNode.textValue()
-            user.url = userNode.textValue()
-
-            val planNode = userNode.get("plan")
-            user.plan = if (planNode.isNull) null else context.readValue(planNode.traverse(), UserPlan::class.java)
-
-            val quarterCommitCount: MutableMap<String, Int> = parseInt(userProfileNode.get("quarterCommitCount"))
-            val langRepoCount: MutableMap<String, Int> = parseInt(userProfileNode.get("langRepoCount"))
-            val langStarCount: MutableMap<String, Int> = parseInt(userProfileNode.get("langStarCount"))
-            val langCommitCount: MutableMap<String, Int> = parseInt(userProfileNode.get("langCommitCount"))
-            val repoCommitCount: MutableMap<String, Int> = parseInt(userProfileNode.get("repoCommitCount"))
-            val repoStarCount: MutableMap<String, Int> = parseInt(userProfileNode.get("repoStarCount"))
-            val repoCommitCountDescriptions: MutableMap<String, String?> = parseString(userProfileNode.get("repoCommitCountDescriptions"))
-            val repoStarCountDescriptions: MutableMap<String, String?> = parseString(userProfileNode.get("repoStarCountDescriptions"))
-
-            return UserProfile(
-                user,
-                quarterCommitCount,
-                langRepoCount,
-                langStarCount,
-                langCommitCount,
-                repoCommitCount,
-                repoStarCount,
-                repoCommitCountDescriptions,
-                repoStarCountDescriptions
-            )
-        }
-
-        private fun parseInt(jsonNode: JsonNode): MutableMap<String, Int> {
-            val map: MutableMap<String, Int> = mutableMapOf()
-            if (!jsonNode.isNull) {
-                val fieldNames = jsonNode.fieldNames()
-                while (fieldNames.hasNext()) {
-                    val key = fieldNames.next()
-                    val value = jsonNode[key].intValue()
-                    map[key] = value
-                }
-            }
-
-            return map
-        }
-
-        private fun parseString(jsonNode: JsonNode): MutableMap<String, String?> {
-            val map: MutableMap<String, String?> = mutableMapOf()
-            if (!jsonNode.isNull) {
-                val fieldNames = jsonNode.fieldNames()
-                while (fieldNames.hasNext()) {
-                    val key = fieldNames.next()
-                    val value = jsonNode[key].textValue()
-                    map[key] = value
-                }
-            }
-
-            return map
-        }
     }
 }
